build: improve integer conversion procedure
The LUT should get placed on its own cache line. This also means
that the look up can be simplified by clamping the converted
unsigned value to the table length and then using the table value
to determine if the digit was valid. This also works for the
base-10 conversion by clamping the test value to be less than 'a'.